requirements.txt
asgiref==3.8.1
Django==5.1.1
pillow==10.4.0
PyJWT==2.9.0
sqlparse==0.5.1
tzdata==2024.2
